+++ work_plan/__manifest__.py
@@ -12,14 +12,13 @@
     'data': [
         "security/groups.xml",
         "security/ir.model.access.csv",
-        "views/toto_class_type.xml",
+        "views/toto_work_line.xml",
         "views/toto_work_product.xml",
-        "views/toto_work_plan.xml",
+        "views/toto_work_device.xml",
+        "views/toto_shaping_plan.xml",
         "views/toto_power_plan.xml",
         "views/toto_dissolve_plan.xml",
-        "views/toto_dissolve_device.xml",
         "views/toto_process_plan.xml",
-        "views/toto_work_device.xml",
         "views/assets.xml",
         "views/menus.xml",
     ],
