+++ work_plan/models/__init__.py
@@ -1,9 +1,8 @@
-from . import toto_class_type
 from . import toto_work_device
 from . import toto_work_plan
 from . import toto_power_plan
-from . import toto_dissolve_device
 from . import toto_dissolve_plan
 from . import toto_process_plan
 from . import toto_work_product
 from . import toto_work_line
+from . import toto_shaping_plan
