+++ work_plan/security/ir.model.access.csv
@@ -1,23 +1,26 @@
 id,name,model_id/id,group_id/id,perm_read,perm_write,perm_create,perm_unlink
 access_toto_work_plan_product_group_user,access_toto_work_plan_product_group_user,model_toto_work_plan_product,group_work_plan_user,1,0,0,0
-access_toto_work_plan_product_group_manager,access_toto_work_plan_product_group_manager,model_toto_work_plan_product,group_work_plan_manager,1,1,1,1
+access_toto_work_plan_product_group_dissolve_monitor,access_toto_work_plan_product_group_dissolve_monitor,model_toto_work_plan_product,group_work_dissolve_monitor,1,1,1,1
 access_toto_work_line_group_user,access_toto_work_line_group_user,model_toto_work_line,group_work_plan_user,1,0,0,0
-access_toto_work_line_group_manager,access_toto_work_line_group_manager,model_toto_work_line,group_work_plan_manager,1,1,1,1
-access_toto_work_class_type_group_user,access_toto_work_class_type_group_user,model_toto_work_class_type,group_work_plan_user,1,0,0,0
-access_toto_work_class_type_group_manager,access_toto_work_class_type_group_manager,model_toto_work_class_type,group_work_plan_manager,1,1,1,1
+access_toto_work_line_group_dissolve_monitor,access_toto_work_line_group_dissolve_monitor,model_toto_work_line,group_work_dissolve_monitor,1,1,1,1
 access_toto_work_device_group_user,access_toto_work_device_group_user,model_toto_work_device,group_work_plan_user,1,0,0,0
-access_toto_work_device_group_manager,access_toto_work_device_group_manager,model_toto_work_device,group_work_plan_manager,1,1,1,1
-access_toto_work_plan_item_group_user,access_toto_work_plan_item_group_user,model_toto_work_plan_item,group_work_plan_user,1,0,0,0
-access_toto_work_plan_item_group_manager,access_toto_work_plan_item_group_manager,model_toto_work_plan_item,group_work_plan_manager,1,1,1,1
-access_toto_work_plan_group_user,access_toto_work_plan_group_user,model_toto_work_plan,group_work_plan_user,1,0,0,0
-access_toto_work_plan_group_manager,access_toto_work_plan_group_manager,model_toto_work_plan,group_work_plan_manager,1,1,1,1
+access_toto_work_device_group_shaping,access_toto_work_device_group_shaping,model_toto_work_device,group_work_shaping_monitor,1,1,1,1
+access_toto_work_device_group_dissolve,access_toto_work_device_group_dissolve,model_toto_work_device,group_work_dissolve_monitor,1,1,1,1
+access_toto_work_device_group_power,access_toto_work_device_group_power,model_toto_work_device,group_work_power_monitor,1,1,1,1
+access_toto_work_device_group_process,access_toto_work_device_group_process,model_toto_work_device,group_work_process_monitor,1,1,1,1
+access_toto_work_plan_item_group_user,access_toto_work_plan_item_group_user,model_toto_work_plan_item,group_work_plan_user,1,1,1,1
+access_toto_work_plan_group_user,access_toto_work_plan_group_user,model_toto_work_plan,group_work_plan_user,1,1,1,1
+access_toto_shaping_plan_group_user,access_toto_shaping_plan_group_user,model_toto_shaping_plan,group_work_plan_user,1,0,0,0
+access_toto_shaping_plan_group_monitor,access_toto_shaping_plan_group_monitor,model_toto_shaping_plan,group_work_shaping_monitor,1,1,1,1
+access_toto_shaping_plan_item_group_user,access_toto_shaping_plan_item_group_user,model_toto_shaping_plan_item,group_work_plan_user,1,0,0,0
+access_toto_shaping_plan_item_group_monitor,access_toto_shaping_plan_item_group_monitor,model_toto_shaping_plan_item,group_work_shaping_monitor,1,1,1,1
 access_toto_power_plan_group_user,access_toto_power_plan_group_user,model_toto_power_plan,group_work_plan_user,1,0,0,0
-access_toto_power_plan_group_manager,access_toto_power_plan_group_manager,model_toto_power_plan,group_work_plan_manager,1,1,1,1
+access_toto_power_plan_group_monitor,access_toto_power_plan_group_monitor,model_toto_power_plan,group_work_power_monitor,1,1,1,1
+access_toto_power_plan_item_group_user,access_toto_power_plan_item_group_user,model_toto_power_plan_item,group_work_plan_user,1,0,0,0
+access_toto_power_plan_item_group_monitor,access_toto_power_plan_item_group_monitor,model_toto_power_plan_item,group_work_power_monitor,1,1,1,1
 access_toto_dissolve_plan_group_user,access_toto_dissolve_plan_group_user,model_toto_dissolve_plan,group_work_plan_user,1,0,0,0
-access_toto_dissolve_plan_group_manager,access_toto_dissolve_plan_group_manager,model_toto_dissolve_plan,group_work_plan_manager,1,1,1,1
+access_toto_dissolve_plan_group_monitor,access_toto_dissolve_plan_group_monitor,model_toto_dissolve_plan,group_work_dissolve_monitor,1,1,1,1
 access_toto_dissolve_plan_item_group_user,access_toto_dissolve_plan_item_group_user,model_toto_dissolve_plan_item,group_work_plan_user,1,0,0,0
-access_toto_dissolve_plan_item_group_manager,access_toto_dissolve_plan_item_group_manager,model_toto_dissolve_plan_item,group_work_plan_manager,1,1,1,1
+access_toto_dissolve_plan_item_group_monitor,access_toto_dissolve_plan_item_group_monitor,model_toto_dissolve_plan_item,group_work_dissolve_monitor,1,1,1,1
 access_toto_process_plan_group_user,access_toto_process_plan_group_user,model_toto_process_plan,group_work_plan_user,1,0,0,0
-access_toto_process_plan_group_manager,access_toto_process_plan_group_manager,model_toto_process_plan,group_work_plan_manager,1,1,1,1
-access_toto_dissolve_device_group_user,access_toto_dissolve_device_group_user,model_toto_dissolve_device,group_work_plan_user,1,0,0,0
-access_toto_dissolve_device_group_manager,access_toto_dissolve_device_group_manager,model_toto_dissolve_device,group_work_plan_manager,1,1,1,1
+access_toto_process_plan_group_monitor,access_toto_process_plan_group_monitor,model_toto_process_plan,group_work_process_monitor,1,1,1,1
