+++ work_plan/views/assets.xml
@@ -2,8 +2,8 @@
 <odoo>
     <template id="assets_backend" name="account assets" inherit_id="web.assets_backend">
         <xpath expr="." position="inside">
-            <script type="text/javascript" src="/work_plan/static/src/js/section_and_note_fields_backend.js"/>
-            <link rel="stylesheet" href="/work_plan/static/src/scss/section_and_note_backend.scss"/>
+<!--            <script type="text/javascript" src="/work_plan/static/src/js/section_and_note_fields_backend.js"/>-->
+<!--            <link rel="stylesheet" href="/work_plan/static/src/scss/section_and_note_backend.scss"/>-->
         </xpath>
     </template>
 </odoo>
