--- work_plan/views/toto_class_type.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="toto_work_class_type_view_form" model="ir.ui.view">
-        <field name="name">toto.work.class.type.view.form</field>
-        <field name="model">toto.work.class.type</field>
-        <field name="arch" type="xml">
-            <form>
-                <sheet>
-                    <group>
-                        <field name="name"/>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record id="toto_work_class_type_view_tree" model="ir.ui.view">
-        <field name="name">toto.work.class.type.view.tree</field>
-        <field name="model">toto.work.class.type</field>
-        <field name="arch" type="xml">
-            <tree editable="bottom">
-                <field name="name"/>
-            </tree>
-        </field>
-    </record>
-
-    <record id="toto_work_class_type_action" model="ir.actions.act_window">
-        <field name="name">班别</field>
-        <field name="type">ir.actions.act_window</field>
-        <field name="res_model">toto.work.class.type</field>
-        <field name="view_mode">tree,form</field>
-        <field name="help" type="html">
-          <p class="oe_view_nocontent_create">
-            <!-- Add Text Here -->
-          </p><p>
-            <!-- More details about what a user can do with this object will be OK -->
-          </p>
-        </field>
-    </record>
-</odoo>
