--- work_plan/views/toto_dissolve_device.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="toto_dissolve_device_view_form" model="ir.ui.view">
-        <field name="name">toto.dissolve.device.view.form</field>
-        <field name="model">toto.dissolve.device</field>
-        <field name="arch" type="xml">
-            <form>
-                <sheet>
-                    <group>
-                        <field name="name"/>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record id="toto_dissolve_device_view_tree" model="ir.ui.view">
-        <field name="name">toto.dissolve.device.view.tree</field>
-        <field name="model">toto.dissolve.device</field>
-        <field name="arch" type="xml">
-            <tree editable="bottom">
-                <field name="name"/>
-            </tree>
-        </field>
-    </record>
-
-    <record id="toto_dissolve_device_action" model="ir.actions.act_window">
-        <field name="name">设备</field>
-        <field name="type">ir.actions.act_window</field>
-        <field name="res_model">toto.dissolve.device</field>
-        <field name="view_mode">tree,form</field>
-        <field name="help" type="html">
-          <p class="oe_view_nocontent_create">
-            <!-- Add Text Here -->
-          </p><p>
-            <!-- More details about what a user can do with this object will be OK -->
-          </p>
-        </field>
-    </record>
-</odoo>
