+++ work_plan/views/toto_power_plan.xml
@@ -1,8 +1,22 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <odoo>
-    <record id="toto_work_plan_view_form_power" model="ir.ui.view">
-        <field name="name">toto_work_plan_view_form_power</field>
-        <field name="model">toto.work.plan</field>
+    <record id="toto_power_plan_view_tree" model="ir.ui.view">
+        <field name="name">toto.power.plan.view.tree</field>
+        <field name="model">toto.power.plan</field>
+        <field name="arch" type="xml">
+            <tree>
+                <field name="date" widget="date"/>
+                <field name="class_type_id"/>
+                <field name="work_type"/>
+                <field name="vacation"/>
+                <field name="staffing"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="toto_power_plan_view_form" model="ir.ui.view">
+        <field name="name">toto_power_plan_view_form</field>
+        <field name="model">toto.power.plan</field>
         <field name="arch" type="xml">
             <form>
                 <field name="plan_type" invisible="1"/>
@@ -31,8 +45,8 @@
                             <field name="actual_quantity"/>
                             <field name="note"/>
                             <field name="state"/>
-                            <button name="action_finish" type="object" string="完成" class="oe_highlight" states="progress"/>
-                            <button name="action_cancel" type="object" string="终止" class="oe_highlight" states="progress,done"/>
+                            <button name="action_finish" type="object" icon="fa-check" states="progress"/>
+                            <button name="action_cancel" type="object" icon="fa-times" states="progress,done"/>
                         </tree>
                         <form>
                             <field name="display_type" invisible="1"/>
@@ -40,7 +54,8 @@
                                 <field name="name" attrs="{'invisible':[('display_type', '!=', 'line_section')]}"/>
                                 <group col="4">
                                     <field name="device_id" attrs="{'invisible':[('display_type', '=', 'line_section')]}" options="{'no_create':True}"/>
-                                    <field name="employee_id" attrs="{'invisible':[('display_type', '=', 'line_section')]}" options="{'no_create':True}"/>
+                                    <field name="employee_id" attrs="{'invisible':[('display_type', '=', 'line_section')]}"
+                                           options="{'no_create':True}" domain="[('department_id', '=', parent.class_type_id)]"/>
                                     <field name="work_subject" attrs="{'invisible':[('display_type', '=', 'line_section')]}"/>
                                     <field name="predetermined_quantity" attrs="{'invisible':[('display_type', '=', 'line_section')]}"/>
                                     <field name="actual_quantity" attrs="{'invisible':[('display_type', '=', 'line_section')]}"/>
@@ -49,11 +64,6 @@
                             </sheet>
                         </form>
                     </field>
-                    <group string="不良对策" attrs="{'invisible':[('plan_type', '=', 'power')]}">
-                        <field name="bad_content"/>
-                        <field name="employee_ids" options="{'no_create':True}" widget="many2many_tags"/>
-                        <field name="detail_countermeasure"/>
-                    </group>
                 </sheet>
                 <div class="oe_chatter">
                     <field name="message_follower_ids" widget="mail_followers"/>
@@ -69,7 +79,7 @@
         <field name="model">toto.work.plan</field>
         <field name="arch" type="xml">
             <calendar string="日历" date_start="date" event_open_popup="true" mode="month"
-                      date_stop="date" quick_add="false" form_view_id="ref('toto_work_plan_view_form_power')">
+                      date_stop="date" quick_add="false" form_view_id="ref('toto_power_plan_view_form')">
             </calendar>
         </field>
     </record>
@@ -77,14 +87,8 @@
     <record id="toto_power_plan_action" model="ir.actions.act_window">
         <field name="name">动力系作业内容</field>
         <field name="type">ir.actions.act_window</field>
-        <field name="res_model">toto.work.plan</field>
-        <field name="context">{'default_plan_type':'power'}</field>
+        <field name="res_model">toto.power.plan</field>
         <field name="view_mode">tree,form,calendar</field>
-        <field name="domain">[('plan_type', '=', 'power')]</field>
-        <field name="view_ids" eval="[(5, 0, 0),
-            (0, 0, {'view_mode': 'calendar', 'view_id': ref('toto_work_plan_view_calendar')}),
-            (0, 0, {'view_mode': 'tree', 'view_id': ref('toto_work_plan_view_tree')}),
-            (0, 0, {'view_mode': 'form', 'view_id': ref('toto_work_plan_view_form_power')})]"/>
         <field name="help" type="html">
             <p class="oe_view_nocontent_create">
                 <!-- Add Text Here -->
