+++ work_plan/views/toto_shaping_plan.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <odoo>
-    <record id="toto_work_plan_view_form" model="ir.ui.view">
-        <field name="name">toto.work.plan.view.form</field>
-        <field name="model">toto.work.plan</field>
+    <record id="toto_shaping_plan_view_form" model="ir.ui.view">
+        <field name="name">toto.shaping.plan.view.form</field>
+        <field name="model">toto.shaping.plan</field>
         <field name="arch" type="xml">
             <form>
                 <field name="plan_type" invisible="1"/>
@@ -62,9 +62,9 @@
         </field>
     </record>
 
-    <record id="toto_work_plan_view_tree" model="ir.ui.view">
-        <field name="name">toto.work.plan.view.tree</field>
-        <field name="model">toto.work.plan</field>
+    <record id="toto_shaping_plan_view_tree" model="ir.ui.view">
+        <field name="name">toto.shaping.plan.view.tree</field>
+        <field name="model">toto.shaping.plan</field>
         <field name="arch" type="xml">
             <tree>
                 <field name="date" widget="date"/>
@@ -79,19 +79,10 @@
     <record id="toto_shaping_plan_action" model="ir.actions.act_window">
         <field name="name">成形系作业内容</field>
         <field name="type">ir.actions.act_window</field>
-        <field name="res_model">toto.work.plan</field>
-        <field name="context">{'default_plan_type':'shaping'}</field>
+        <field name="res_model">toto.shaping.plan</field>
         <field name="view_mode">tree,form</field>
-        <field name="domain">[('plan_type', '=', 'shaping')]</field>
-        <field name="view_ids" eval="[(5, 0, 0),
-            (0, 0, {'view_mode': 'tree', 'view_id': ref('toto_work_plan_view_tree')}),
-            (0, 0, {'view_mode': 'form', 'view_id': ref('toto_work_plan_view_form')})]"/>
         <field name="help" type="html">
             <p class="oe_view_nocontent_create">
-                <!-- Add Text Here -->
-            </p>
-            <p>
-                <!-- More details about what a user can do with this object will be OK -->
             </p>
         </field>
     </record>
